+++ core/src/components/menu/test/focus-trap/e2e.ts
@@ -12,7 +12,7 @@ test('menu: focus trap with overlays', async () => {
 
   const ionDidOpen = await page.spyOnEvent('ionDidOpen');
   const ionModalDidPresent = await page.spyOnEvent('ionModalDidPresent');
-  const ionModalDidDismiss= await page.spyOnEvent('ionModalDidDismiss');
+  const ionModalDidDismiss = await page.spyOnEvent('ionModalDidDismiss');
 
   const menu = await page.find('ion-menu');
   await menu.callMethod('open');
@@ -40,7 +40,6 @@ test('menu: focus trap with content inside overlays', async () => {
 
   const ionDidOpen = await page.spyOnEvent('ionDidOpen');
   const ionModalDidPresent = await page.spyOnEvent('ionModalDidPresent');
-  const ionModalDidDismiss= await page.spyOnEvent('ionModalDidDismiss');
 
   const menu = await page.find('ion-menu');
   await menu.callMethod('open');
@@ -56,4 +55,37 @@ test('menu: focus trap with content inside overlays', async () => {
   await button.click();
 
   expect(await getActiveElementID(page)).toEqual('other-button');
+});
+
+test('menu: should work with swipe gestures after modal is dismissed', async () => {
+  const page = await newE2EPage({
+    url: '/src/components/menu/test/focus-trap?ionic:_testing=true'
+  });
+
+  const ionDidOpen = await page.spyOnEvent('ionDidOpen');
+  const ionModalDidPresent = await page.spyOnEvent('ionModalDidPresent');
+  const ionModalDidDismiss = await page.spyOnEvent('ionModalDidDismiss');
+
+  const menu = await page.find('ion-menu');
+  await menu.callMethod('open');
+  await ionDidOpen.next();
+
+  const openModal = await page.find('#open-modal-button');
+  await openModal.click();
+  await ionModalDidPresent.next();
+
+  const modal = await page.find('ion-modal');
+  await modal.callMethod('dismiss');
+  await ionModalDidDismiss.next();
+
+  await page.mouse.move(30, 168);
+  await page.mouse.down();
+
+  await page.mouse.move(384, 168);
+  await page.mouse.up();
+
+  await page.waitForChanges();
+
+  expect(menu.classList.contains('show-menu')).toBeTruthy();
+
 });