+++ core/stencil.config.ts
@@ -261,6 +261,9 @@ export const config: Config = {
     allowableMismatchedPixels: 200,
     pixelmatchThreshold: 0.05,
     waitBeforeScreenshot: 20,
+    moduleNameMapper: {
+      "@utils/test": ["<rootDir>/src/utils/test/utils"]
+    },
     emulate: [
       {
         userAgent: 'iPhone',